chore: move back to agent package
