+++ packages/agent/src/harness/session/sqlite/index.ts
