+++ packages/agent/src/harness/session/sqlite/storage/session-entries.ts
@@ -1,4 +1,4 @@
-import type { SessionTreeEntry, SessionTreeEntryBase } from "@earendil-works/pi-agent-core";
+import type { SessionTreeEntry, SessionTreeEntryBase } from "../../../types.ts";
 import { invalidEntry, isRecord } from "./shared.ts";
 
 export interface SessionEntryRow {
