+++ packages/agent/tsconfig.build.json
@@ -9,5 +9,5 @@
 		"rootDir": "./src"
 	},
 	"include": ["src/**/*.ts"],
-	"exclude": ["node_modules", "dist", "**/*.d.ts", "src/**/*.d.ts", "src/harness/session/sqlite/**"]
+	"exclude": ["node_modules", "dist", "**/*.d.ts", "src/**/*.d.ts"]
 }
