--- packages/session-backend-sqlite/package.json
@@ -1,70 +1,0 @@
-{
-	"name": "@earendil-works/pi-session-backend-sqlite",
-	"version": "0.80.6",
-	"description": "SQLite session backend for pi agent sessions",
-	"type": "module",
-	"main": "./dist/index.js",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.js"
-		},
-		"./env/node": {
-			"types": "./dist/env/node.d.ts",
-			"import": "./dist/env/node.js"
-		},
-		"./env/bun": {
-			"types": "./dist/env/bun.d.ts",
-			"import": "./dist/env/bun.js"
-		},
-		"./env/better-sqlite3": {
-			"types": "./dist/env/better-sqlite3.d.ts",
-			"import": "./dist/env/better-sqlite3.js"
-		},
-		"./env/sqlite3": {
-			"types": "./dist/env/sqlite3.d.ts",
-			"import": "./dist/env/sqlite3.js"
-		},
-		"./package.json": "./package.json"
-	},
-	"files": [
-		"dist",
-		"sql",
-		"README.md"
-	],
-	"scripts": {
-		"clean": "shx rm -rf dist",
-		"build": "tsgo -p tsconfig.build.json",
-		"prepublishOnly": "npm run clean && npm run build"
-	},
-	"dependencies": {
-		"@earendil-works/pi-agent-core": "^0.80.6"
-	},
-	"optionalDependencies": {
-		"better-sqlite3": "12.11.1",
-		"sqlite3": "6.0.1"
-	},
-	"devDependencies": {
-		"@types/better-sqlite3": "7.6.13",
-		"@types/node": "24.12.4",
-		"shx": "0.4.0",
-		"typescript": "5.9.3"
-	},
-	"keywords": [
-		"pi",
-		"agent",
-		"sqlite",
-		"session"
-	],
-	"author": "Earendil Works",
-	"license": "MIT",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/earendil-works/pi.git",
-		"directory": "packages/session-backend-sqlite"
-	},
-	"engines": {
-		"node": ">=22.19.0"
-	}
-}
