--- packages/session-backend-sqlite/tsconfig.build.json
@@ -1,14 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "./dist",
-		"paths": {
-			"@earendil-works/pi-agent-core": ["../agent/dist/index.d.ts"],
-			"@earendil-works/pi-agent-core/node": ["../agent/dist/node.d.ts"],
-			"@earendil-works/pi-agent-core/*": ["../agent/dist/*.d.ts"]
-		},
-		"rootDir": "./src"
-	},
-	"include": ["src/**/*.ts"],
-	"exclude": ["node_modules", "dist", "**/*.d.ts", "src/**/*.d.ts"]
-}
