+++ tsconfig.json
@@ -11,14 +11,17 @@
 			"@earendil-works/pi-ai/*": ["./packages/ai/src/*.ts", "./packages/ai/src/providers/*.ts"],
 			"@earendil-works/pi-ai/dist/*": ["./packages/ai/src/*"],
 			"@earendil-works/pi-agent-core": ["./packages/agent/src/index.ts"],
+			"@earendil-works/pi-agent-core/sqlite": ["./packages/agent/src/harness/session/sqlite/index.ts"],
+			"@earendil-works/pi-agent-core/sqlite/env/node": ["./packages/agent/src/harness/session/sqlite/env/node.ts"],
+			"@earendil-works/pi-agent-core/sqlite/env/bun": ["./packages/agent/src/harness/session/sqlite/env/bun.ts"],
+			"@earendil-works/pi-agent-core/sqlite/env/better-sqlite3": ["./packages/agent/src/harness/session/sqlite/env/better-sqlite3.ts"],
+			"@earendil-works/pi-agent-core/sqlite/env/sqlite3": ["./packages/agent/src/harness/session/sqlite/env/sqlite3.ts"],
 			"@earendil-works/pi-agent-core/*": ["./packages/agent/src/*"],
 			"@earendil-works/pi-coding-agent": ["./packages/coding-agent/src/index.ts"],
 			"@earendil-works/pi-coding-agent/hooks": ["./packages/coding-agent/src/core/hooks/index.ts"],
 			"@earendil-works/pi-coding-agent/*": ["./packages/coding-agent/src/*"],
 			"@earendil-works/pi-orchestrator": ["./packages/orchestrator/src/index.ts"],
 			"@earendil-works/pi-orchestrator/*": ["./packages/orchestrator/src/*"],
-			"@earendil-works/pi-session-backend-sqlite": ["./packages/session-backend-sqlite/src/index.ts"],
-			"@earendil-works/pi-session-backend-sqlite/*": ["./packages/session-backend-sqlite/src/*"],
 			"typebox": ["./node_modules/typebox"],
 			"@earendil-works/pi-tui": ["./packages/tui/src/index.ts"],
 			"@earendil-works/pi-tui/*": ["./packages/tui/src/*"],
@@ -27,9 +30,5 @@
 		}
 	},
 	"include": ["packages/*/src/**/*", "packages/*/test/**/*", "packages/coding-agent/examples/**/*"],
-	"exclude": [
-		"**/dist/**",
-		"packages/agent/src/harness/session/sqlite/**",
-		"packages/coding-agent/examples/extensions/gondolin/**"
-	]
+	"exclude": ["**/dist/**", "packages/coding-agent/examples/extensions/gondolin/**"]
 }
